refactor(hooks): extract attachment hook outside UI components
